Refactors constants and SSL generation methods from VaultContainer to TestConstants and SSLUtils
