+++ src/test-integration/java/com/bettercloud/vault/api/AuthBackendAppIdTests.java
@@ -2,6 +2,7 @@ package com.bettercloud.vault.api;
 
 import com.bettercloud.vault.Vault;
 import com.bettercloud.vault.VaultException;
+import com.bettercloud.vault.util.VaultContainer;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
 import org.junit.Test;
