+++ src/test-integration/java/com/bettercloud/vault/api/DebugTests.java
@@ -3,6 +3,7 @@ package com.bettercloud.vault.api;
 import com.bettercloud.vault.Vault;
 import com.bettercloud.vault.VaultException;
 import com.bettercloud.vault.response.HealthResponse;
+import com.bettercloud.vault.util.VaultContainer;
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
